Python types for CommandFinished
